refactor: extract profile commands from main.rs to dedicated module
- Create commands/profile.rs with all profile command handlers
- Move 660 lines of profile logic out of main.rs
- Reduce main.rs from 1,589 lines to 929 lines (41.5% reduction)
- Split into focused handlers: list, path, show, set, remove, validate, default-enterprise, default-cloud
- Update commands/mod.rs to include profile module
- Remove unused Context import from main.rs
- All 45 tests passing
- No clippy warnings

Main.rs now focuses on CLI parsing, initialization, and command routing.

Closes #402
